[Implement]: 원시값 포장을 위해 Number생성, 다른코드 임시원복
