+++ src/main/java/step3/InputView.java
@@ -21,14 +21,9 @@ public class InputView {
         return inputWinningNumbers;
     }
 
-    public static String inputWinningNumbers2() {
-        System.out.println("지난 주 당첨 번호를 입력해 주세요.");
-        return scan.nextLine();
-    }
-
     public static void bonusNumberAdd(Lotto inputWinningLottoNumberList) {
         System.out.println("보너스 볼을 입력해주세요.");
         Lotto.lottoBonusNumber = scan.nextInt();
-        LottoNo.checkBonusNumber(inputWinningLottoNumberList);
+        LottoNumber.checkBonusNumber(inputWinningLottoNumberList);
     }
 }
