+++ src/main/java/step3/Number.java
@@ -1,0 +1,4 @@
+package step3;
+
+public class Number {
+}
